# internal/docs/swagger.yaml
definitions:
...
  github_com_umar5678_go-backend_internal_modules_homeservices_dto.AddOnResponse:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      durationMinutes:
        type: integer
      id:
        type: integer
      imageUrl:
        type: string
      isActive:
        type: boolean
      originalPrice:
        type: number
      price:
        type: number
      sortOrder:
        type: integer
      title:
        type: string
    type: object
...
  github_com_umar5678_go-backend_internal_modules_homeservices_dto.CreateCategoryRequest:
    properties:
      bannerImage:
        type: string
      description:
        maxLength: 1000
        type: string
      highlights:
        items:
          type: string
        type: array
      iconUrl:
        type: string
      isActive:
        type: boolean
      name:
        maxLength: 150
        minLength: 2
        type: string
      sortOrder:
        type: integer
    required:
    - description
    - name
    type: object
...
  github_com_umar5678_go-backend_internal_modules_homeservices_dto.ServiceListResponse:
    properties:
      basePrice:
        type: number
      categoryId:
        type: integer
      createdAt:
        type: string
      discountPercentage:
        type: integer
      durationMinutes:
        type: integer
      id:
        type: integer
      imageUrl:
        type: string
      isActive:
        type: boolean
      isFeatured:
        type: boolean
      name:
        type: string
      originalPrice:
        type: number
      tabId:
        type: integer
    type: object
...
  github_com_umar5678_go-backend_internal_utils_response.Response:
    properties:
      code:
        type: string
      data: {}
      errors:
        description: Use ErrorDetail
        items:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.ErrorDetail'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Meta'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-grade Go backend API
  title: supr booking server in go
  version: "1.0"
paths:
...
  /services:
    get:
      description: Get paginated list of services with filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category ID
        in: query
        name: categoryId
        type: integer
      - description: Search in name and description
        in: query
        name: search
        type: string
      - description: Minimum price
        in: query
        name: minPrice
        type: number
      - description: Maximum price
        in: query
        name: maxPrice
        type: number
      - description: Filter by active status
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.ServiceListResponse'
                  type: array
              type: object
      summary: List services
      tags:
      - home-services
  /services/{id}:
    get:
      description: Get detailed service information including options
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.ServiceDetailResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      summary: Get service details
      tags:
      - home-services
  /services/addons:
    get:
      description: Get all available add-ons for a category
      parameters:
      - description: Category ID
        in: query
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.AddOnResponse'
                  type: array
              type: object
      summary: List add-ons
      tags:
      - home-services
  /services/admin/addons:
    post:
      consumes:
      - application/json
      description: Create a new service add-on (admin only)
      parameters:
      - description: Add-on details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.CreateAddOnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.AddOnResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create an add-on
      tags:
      - home-services-admin
  /services/admin/categories:
    post:
      consumes:
      - application/json
      description: Create a new service category (admin only)
      parameters:
      - description: Category details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.CategoryWithTabsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - home-services-admin
  /services/admin/services:
    post:
      consumes:
      - application/json
      description: Create a new home service (admin only)
      parameters:
      - description: Service details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.ServiceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Create a service
      tags:
      - home-services-admin
  /services/admin/services/{id}:
    put:
      consumes:
      - application/json
      description: Update service details (admin only)
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.ServiceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Update a service
      tags:
      - home-services-admin
  /services/admin/tabs:
    post:
      consumes:
      - application/json
      description: Create a new service tab/subcategory (admin only)
      parameters:
      - description: Tab details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.CreateTabRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.ServiceTabResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a tab
      tags:
      - home-services-admin
...
  /services/categories/{id}:
    get:
      description: Get detailed category information with all tabs
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.CategoryWithTabsResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      summary: Get category with tabs
      tags:
      - home-services
  /services/orders:
    get:
      description: Get authenticated user's service orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.OrderListResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Get my orders
      tags:
      - home-services
    post:
      consumes:
      - application/json
      description: Book a home service
      parameters:
      - description: Order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Create a service order
      tags:
      - home-services
  /services/orders/{id}:
    get:
      description: Get detailed information about a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.OrderResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Get order details
      tags:
      - home-services
  /services/orders/{id}/cancel:
    post:
      description: Cancel a pending or searching order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - home-services
  /services/provider/orders:
    get:
      description: Get orders assigned to the provider
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_homeservices_dto.OrderListResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Get provider orders
      tags:
      - home-services-provider
  /services/provider/orders/{id}/accept:
    post:
      description: Accept a job offer
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Accept an order
      tags:
      - home-services-provider
...
swagger: "2.0"